Use generator for get_all_serial
get_all_serial is used to iterate through a paginated response from
shipwire's API. Previously it would add each batch of items to a list
and then return that list once there were no more results to get.

This changes get_all_serial to be a generator, so that the program only
has to keep a single batch's worth of items in memory at a time. This is
especially helpful for grabbing large datasets from shipwire, because
you can begin processing items from shipwire while there are still items
to be requested.